src/parsing/requirements.txt
kafka-python==2.1.5
PyXB-X==1.2.6.3
